// client/src/app/models/IProduct.ts
export interface IProduct {
    id: number;
    name: string;
    description: string;
    price: number;
    brand: string;
    model: string;
    year: number;
    os: string;
    color: string;
    size: number;
    resolution: string;
    camera: number;
    battery: number;
    processor: string;
    memory: number;
    storage: number;
    imageUrl: string;
    ownerUsername: string;
    dateAdded: string;
}